requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.2
idna==3.2
python-dotenv==0.18.0
requests==2.26.0
selenium==3.141.0
six==1.16.0
spotipy==2.18.0
urllib3==1.26.6
